# web-poll-swagger/api-gateway/api/docs/swagger.yaml
definitions:
  genprotos.ByEmail:
    properties:
      email:
        type: string
    type: object
...
  models.ForgotPasswordReq:
    properties:
      email:
        description: User's email address
        type: string
    type: object
...
  models.RecoverPasswordReq:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a confirmation code to email recovery password
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Page not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Forgot passwrod
      tags:
      - password-recovery
...
  /recover-password:
    post:
      consumes:
      - application/json
      description: Verifies the code and updates the password
      parameters:
      - description: Recover Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RecoverPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Password successfully updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Incorrect verification code
          schema:
            type: string
        "404":
          description: Verification code expired or email not found
          schema:
            type: string
        "500":
          description: Error updating password
          schema:
            type: string
      summary: Recover password (Use this one after sending verification code)
      tags:
      - password-recovery
...
  /resverify:
    post:
      consumes:
      - application/json
      description: Sends a confirmation code to email recovery password again
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/genprotos.ByEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Page not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Sends code again if user didn't recieve the code
      tags:
      - password-recovery
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"